Chore: move menu to main
