+++ main.rb
@@ -1,5 +1,4 @@
 require_relative 'app'
-require_relative 'menu_choice'
 
 def options_menu
   puts 'Please choose your menu option!'
@@ -12,6 +11,17 @@ def options_menu
   puts '6 - List all rentals for person(id)'
   puts '7 - Exit'
   puts "\n"
+end
+
+def handle_menu_choice(choice, app)
+  case choice
+  when 1 then app.list_all_books
+  when 2 then app.list_all_people
+  when 3 then app.create_person
+  when 4 then app.create_a_book
+  when 5 then app.create_a_rental
+  when 6 then app.rental_person_id
+  end
 end
 
 def main
